Proposed breaking change to Attendee deserialization for v2
Currently the deserialization of an attendee will return
null if the calendar user address is not a valid URI, even
after some custom fixes. I propose we change this to always
return an attendee. Also, an additive change to include
the original calendar user address in a property of Attendee.