Remove net46 framework from net-core
When this is present the build will break on the mac b/c it doesn't have .net 4.6 installed.  Instead of trying to work around that (which isn't needed since we want netstandard for .net core support), we instead deprecate it.

We didn't go so far as to remove the .Net framework dirs as that would result in lots of git merge issues if we take new drops from master.