Fix weekly bymonth
Weekly rules that include BYMONTH were only including weeks that
start inside the given months. This fix checks the end of the week
also to see if any part of the week is inside the given months.

The date will not always be the start of a week, since the
reference date is first, but extra months are filtered out later.

Fixes #879