Zoned occurrence using NodaTime (#854)
* Use BCL HashCode

* Fix date parsing in a test

* Evaluate recurrence with nodatime

* Evaluate calendar in a time zone always

This evaluates events in the calendar for a specific time zone.
Events with a time zone are evaluated in that event's time zone and then
converted to the requested time zone. Floating events are evaluated
in the request time zone. This is essentially the same behavior as before,
except the results are actual point-in-time values.

A key evaluation difference is that during backward daylight saving,
the last time offset can be output instead of only the first. Before,
an hourly repeating event would skip the second hour.

Fixes #681, #875

* Change tests to fit zoned API

The few tests still failing were left alone on purpose. With some tests,
the test data needs to be changed because it does not make sense now that
all occurrences are in a specific time zone.

For FREEBUSY tests, that was mostly ignored because I am not sure how it
should be used.

* Remove invalid comparison operators

The only valid comparison for two CalDateTime values is equality.
The greater and less than operators cannot always return a valid
result. This makes implementing IComparable also invalid.
"20250815" != "20250815T000000" but these operators also did not
say they were less or greater either.

* Remove copying of CalDateTime

CalDateTime is immutable, so there is no need to copy.

* Add backward daylight saving tests

* Change benchmarks to fit zoned API

* Store CalDateTime values with NodaTime

* Reduce evaluation allocation

* Fix CalDateTime format test

* Validate CalDateTime year

CalDateTime was validating year using DateTime limits. NodaTime
supports more than just positive years, so verify that the year
is positive to match DateTime behavior.

* Change out of range test

This test had an effective end that exceeded the valid date
range but only actually held a duration, so there was no
exception. The new occurrence format now produces an end
date always, which now exceeds the valid date range.

* Change time zone test

This was essentially testing in the system time zone before.
Now that evaluation requires a time zone, this required
adjusting the local time to fit the default testing time zone
of US Eastern.

* Removed warning

It makes sense to be a nominal duration in this case,
even if the spec does not say so specifically.

* Validate EvaluationPeriod end

* Fix freebusy validation message

* Simplify CalDateTime LocalDateTime conversion

* Change dateonly timeonly to methods

* Remove unused DateUtil methods

* Fix some sonar issues